% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_make_filename}
\alias{fars_make_filename}
\title{Create filename following the Coursera FARS datasets naming convention}
\usage{
fars_make_filename(year = "2013")
}
\arguments{
\item{year}{A character string (or numeric/integer value) specifying year of interest for FARS data.
The function uses as.integer(year), therefore, given agrument should be coercible to integer
Parsing from numeric should be taken with care, 2015.7 will result in 2015.}
}
\value{
This function returns string with filename matching the filename from Coursera dataset (does not include path)
}
\description{
This is a simple function returning the name of the file, according to the year given in the function argument
}
\details{
The function can malfunction in case the specified year is not coercible to integer or database for resulting integer
is missing
}
\examples{
fars_make_filename("2013")
fars_make_filename(2014)
fars_make_filename(2015)

}
