% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_map_state}
\alias{fars_map_state}
\title{Plot FARS accidents on the map of the state (of given ID number)}
\usage{
fars_map_state(state.num, year)
}
\arguments{
\item{state.num}{State identification number, allowed values in range from 1 to 56 without {3,7,14,43,52}}

\item{year}{Year for which the plot should be prepared, provided datasets allow year to be {2013,2014,2015}}
}
\value{
This function does not return any data, side effect of the function is the plot of FARS accident
}
\description{
This function for given state and year plots the map of the FARS accident on the border of the state
This function does not return any data, side effect of the function is the plot of FARS accident on the border map of the state
}
\details{
The function can malfunction in case the dataset for specified year is missing, or is given in wrong format
Additionally, the function can malfunction if the state identifier is not matching the format from dataset for specified year
}
\examples{
fars_map_state(1, 2013)
fars_map_state(12, 2014)
fars_map_state(23, 2013)

}
