% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read}
\alias{fars_read}
\title{Read FARS data}
\usage{
fars_read(filename = farsReport:::fars_make_filename())
}
\arguments{
\item{filename}{A relative file to csv file, with dec="." and sep="," (due to use of read_csv)}
}
\value{
This function returns a tibble, from a specified file
}
\description{
This is a simple function, which reads FARS data from CSV file, specified in the function argument
The function checks for presence of specified file, if present, reads the file and converts to a tibble
}
\details{
The function cannot unpredictably malfunction, all error cases are covered with the if statement
}
\examples{
fars_read(farsReport:::fars_make_filename())

}
