% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read_years}
\alias{fars_read_years}
\title{Load FARS accidents for given years}
\usage{
fars_read_years(years)
}
\arguments{
\item{years}{A list of years, format of each year must be corcible to a valid year by as.integer() function}
}
\value{
This function returns a list of tibbles, in each tibble month and year of each accident will be listed
}
\description{
For given list of years, load the FARS accidents dataset and return as a tibble (per year)
}
\details{
The function can malfunction in case the dataset for specified year is missing, or is given in wrong format
}
\examples{
fars_read_years("2013")
fars_read_years(c(2013, 2014))
fars_read_years(c("2013", "2014", "2015"))

}
