% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{make_filename}
\alias{make_filename}
\title{Generate File name from the US National Highway Traffic Safety
Administration's Fatality Analysis Reporting System
\href{http://www.nhtsa.gov/research-data/fatality-analysis-reporting-system-fars}{Fatality Analysis Reporting System (FARS)}.
This function generates the accident file name for a particular year taken as input. Entering other
than integer may throw a warning}
\usage{
make_filename(year)
}
\arguments{
\item{year}{a integer representing the year for which the file name to be generated}
}
\value{
this function returns the  accident file name for a particular year given as input

 @export

 @examples
 \dontrun{
 make_filename(1992)
 make_filename("2012")
 }
}
\description{
Generate File name from the US National Highway Traffic Safety
Administration's Fatality Analysis Reporting System
\href{http://www.nhtsa.gov/research-data/fatality-analysis-reporting-system-fars}{Fatality Analysis Reporting System (FARS)}.
This function generates the accident file name for a particular year taken as input. Entering other
than integer may throw a warning
}

