% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_map_state}
\alias{fars_map_state}
\title{Plot Accidents for a State}
\usage{
fars_map_state(state.num, year)
}
\arguments{
\item{state.num}{a numeric code for a state(e.g 5 or 6)}

\item{year}{accident year}
}
\value{
a map plotting the accidents for the given state in an input year
}
\description{
This function plot the accidents for a state given as a input
}
\section{Warning}{

Dont enter the invalid state or year,else function will generate an error.
}
\examples{
\dontrun{
fars_map_state(1,2015)
fars_map_state(6,2013)
fars_map_state(2,2013) #2 is invalid state
}

}

