% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read}
\alias{fars_read}
\title{Reads a  CSV File}
\usage{
fars_read(filename)
}
\arguments{
\item{filename}{a charater string giving the file that function will read}
}
\value{
this function returns the data frame
}
\description{
This is a simple function that takes a file name as a input , check for existence,
if yes read that csv file and convert it into a data frame. If file doesnot exist, it will
return an error.
}
\examples{
\dontrun{
fars_read("student_data.csv")
fars_read("accident_2013.csv.bz2")
}

}

