% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read_years}
\alias{fars_read_years}
\title{Read Accident files for years}
\usage{
fars_read_years(years)
}
\arguments{
\item{years}{a numeric vector or list for which files need to be read}
}
\value{
a data frame containing Month and year of the file for each year, else a warning message
for invalid year
}
\description{
This function read files for series of  years and print Month and year column of the data frame
}
\examples{
\dontrun{
fars_read_years(2013)
fars_read_years(2014)
fars_read_years(1990)
}


}

