% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_summarize_years}
\alias{fars_summarize_years}
\title{Summarize Accident over years}
\usage{
fars_summarize_years(years)
}
\arguments{
\item{years}{a numeric vector or list for which files need to be read}
}
\value{
returns a key value pair of year(key) and the total number of accidents(value) in that year
}
\description{
This function summarize accidents over many years and return a key value pair for years and
number of accidents in that year.Throws an error for invalid state number
}
\examples{
\dontrun{
fars_summarize_years(c(2013,2014))
fars_summarize_years(2013)
}


}

